requirements.txt
matplotlib
torch==2.0.0
cvzone==1.5.6
hydra-core>=1.2.0
matplotlib>=3.2.2
numpy==1.24.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
filterpy==1.4.5
scikit-image==0.19.3
opencv-python==4.6.0.66
ultralytics==8.0.26
transformers==4.38.1
easyocr==1.7.0
pandas==1.5.3
paddleocr==2.7.0.3
paddlepaddle
mysql-connector-python

